Added animation encoder error handling
